design and bag-fix

+++ database/tables/tariff.py
@@ -1,4 +1,4 @@
-from peewee import CharField, IntegerField, ForeignKeyField, DateTimeField
+from peewee import CharField, IntegerField, ForeignKeyField, DateTimeField, BigIntegerField
 
 from database.db_connect import BaseModel
 from database.tables.seller import Seller
@@ -7,9 +7,9 @@ from database.tables.seller import Seller
 class Tariff(BaseModel):
     '''Таблица тафрифов'''
     name = CharField(unique=True)
-    price = IntegerField()
-    duration_time = IntegerField() # days
-    feedback_amount = IntegerField()
+    price = BigIntegerField()
+    duration_time = BigIntegerField() # days
+    feedback_amount = BigIntegerField()
 
     class Meta:
         db_table = 'Тарифы'
@@ -19,7 +19,7 @@ class TariffsToSellers(BaseModel):
     tariff = ForeignKeyField(Tariff, backref='tariffs')
     start_date_time = DateTimeField()
     end_date_time = DateTimeField()
-    residual_feedback = IntegerField()
+    residual_feedback = BigIntegerField()
 
     class Meta:
         db_table = 'Тарифы_Продавцы'
